requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
Werkzeug==3.0.1
bcrypt==4.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
vaderSentiment==3.3.2
Pillow==10.1.0
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
